# frontend/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/liquidity/setup-position": {
      "post": {
        "description": "Создает новую позицию в пуле Raydium CLMM с указанным диапазоном цен",
        "operationId": "LiquidityBotController_setupLiquidityPosition",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetupPositionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Позиция успешно создана",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "mint": "5HauEoJa6tcKBy1vofNAxnEDiMpC7H6V9L5TgLwFYNfd",
                    "txId": "https://explorer.solana.com/tx/...",
                    "poolId": "2QdhepnKRTLjjSqPL1PtKNwqrUkoLee5Gqs8bvZhRdMv"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Неверные параметры"
          },
          "401": {
            "description": "Не авторизован"
          },
          "500": {
            "description": "Ошибка сервера"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Открыть позицию ликвидности",
        "tags": [
          "liquidity"
        ]
      }
    },
    "/liquidity/close-position": {
      "post": {
        "description": "Закрывает существующую позицию и выводит все токены",
        "operationId": "LiquidityBotController_closeLiquidityPosition",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClosePositionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Позиция успешно закрыта",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "txId": "https://explorer.solana.com/tx/...",
                    "success": true,
                    "baseAmount": 21.57462,
                    "quoteAmount": 8.705166,
                    "feesCollected": 17.65
                  }
                }
              }
            }
          },
          "404": {
            "description": "Позиция не найдена"
          },
          "500": {
            "description": "Ошибка сервера"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Закрыть позицию ликвидности",
        "tags": [
          "liquidity"
        ]
      }
    },
    "/liquidity/position/{nftMint}": {
      "get": {
        "description": "Возвращает детальную информацию о конкретной позиции",
        "operationId": "LiquidityBotController_getPositionInfo",
        "parameters": [
          {
            "name": "nftMint",
            "required": true,
            "in": "path",
            "description": "NFT Mint адрес позиции",
            "schema": {
              "example": "5HauEoJa6tcKBy1vofNAxnEDiMpC7H6V9L5TgLwFYNfd",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Информация о позиции",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "position": {
                      "positionId": "5HauEoJa...",
                      "baseAmount": "21.526289",
                      "quoteAmount": "0.000000",
                      "priceRange": {
                        "lower": 154.63,
                        "upper": 161.81
                      },
                      "currentPrice": 162.27,
                      "profitability": 5.23
                    },
                    "pool": {
                      "poolId": "2Qdhep...",
                      "baseMint": "SOL",
                      "quoteMint": "USDC",
                      "currentPrice": 162.27
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Позиция не найдена"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Получить информацию о позиции",
        "tags": [
          "liquidity"
        ]
      }
    },
    "/liquidity/positions": {
      "get": {
        "description": "Возвращает список всех активных позиций пользователя",
        "operationId": "LiquidityBotController_getAllPositionInfo",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Список позиций",
            "content": {
              "application/json": {
                "schema": {
                  "example": [
                    {
                      "position": {
                        "positionId": "5HauEoJa...",
                        "baseAmount": "21.526289",
                        "quoteAmount": "0.000000"
                      },
                      "pool": {
                        "poolId": "2Qdhep...",
                        "baseMint": "SOL",
                        "quoteMint": "USDC"
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Получить все позиции",
        "tags": [
          "liquidity"
        ]
      }
    },
    "/liquidity/positions/wallet": {
      "get": {
        "description": "Возвращает все позиции из подключенного кошелька",
        "operationId": "LiquidityBotController_getAllPositionsFromWallet",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Список позиций"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Получить все позиции кошелька",
        "tags": [
          "liquidity"
        ]
      }
    },
    "/liquidity/pool/{poolId}": {
      "get": {
        "description": "Возвращает детали пула ликвидности",
        "operationId": "LiquidityBotController_getPoolInfo",
        "parameters": [
          {
            "name": "poolId",
            "required": true,
            "in": "path",
            "description": "ID пула",
            "schema": {
              "example": "2QdhepnKRTLjjSqPL1PtKNwqrUkoLee5Gqs8bvZhRdMv",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Информация о пуле",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "id": "2Qdhep...",
                    "mintA": {
                      "symbol": "SOL",
                      "decimals": 9
                    },
                    "mintB": {
                      "symbol": "USDC",
                      "decimals": 6
                    },
                    "price": 162.27,
                    "tvl": 5234567.89
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Получить информацию о пуле",
        "tags": [
          "liquidity"
        ]
      }
    },
    "/liquidity/token/price": {
      "get": {
        "description": "Возвращает текущие цены указанных токенов в USD",
        "operationId": "LiquidityBotController_getTokenPrice",
        "parameters": [
          {
            "name": "symbols",
            "required": true,
            "in": "query",
            "description": "Символы токенов через запятую",
            "schema": {
              "example": "SOL,USDC,USDT",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Цены токенов",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "SOL": 162.27,
                    "USDC": 1,
                    "USDT": 1
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Получить цены токенов",
        "tags": [
          "liquidity"
        ]
      }
    },
    "/liquidity/pool/balance/{poolId}": {
      "get": {
        "description": "Возвращает текущие балансы токенов в указанном пуле",
        "operationId": "LiquidityBotController_getPoolBalance",
        "parameters": [
          {
            "name": "poolId",
            "required": true,
            "in": "path",
            "description": "ID пула",
            "schema": {
              "example": "2QdhepnKRTLjjSqPL1PtKNwqrUkoLee5Gqs8bvZhRdMv",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Балансы токенов",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "SOL": {
                      "amount": 21.526289,
                      "symbol": "SOL"
                    },
                    "USDC": {
                      "amount": 3842.701244,
                      "symbol": "USDC"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Получить балансы токенов пула",
        "tags": [
          "liquidity"
        ]
      }
    },
    "/liquidity/increase": {
      "post": {
        "operationId": "LiquidityBotController_increaseLiquidity",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IncreasePositionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "liquidity"
        ]
      }
    },
    "/liquidity/density": {
      "get": {
        "operationId": "LiquidityBotController_getDensityForPosition",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "liquidity"
        ]
      }
    },
...
    "/liquidity/optimal-range-volume/position/{positionId}": {
      "get": {
        "operationId": "OptimalRangeVolumeController_getForPosition",
        "parameters": [
          {
            "name": "positionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "period",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "OptimalRangeVolume"
        ]
      }
    },
    "/liquidity/optimal-range-volume/all": {
      "get": {
        "operationId": "OptimalRangeVolumeController_getForAllPositions",
        "parameters": [
          {
            "name": "period",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "OptimalRangeVolume"
        ]
      }
    }
  },
  "info": {
    "title": "Liquidity Bot API",
    "description": "API для управления ликвидностью на Raydium CLMM",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "SetupPositionDto": {
        "type": "object",
        "properties": {
          "poolId": {
            "type": "string",
            "description": "ID пула",
            "example": "3ucNos4NbumPLZNWztqGHNFFgkHeRMBQAVemeeomsUxv"
          },
          "baseMint": {
            "type": "string",
            "description": "Адрес базового токена (mint)",
            "example": "So11111111111111111111111111111111111111112"
          },
          "quoteMint": {
            "type": "string",
            "description": "Адрес котируемого токена (mint)",
            "example": "EPjFWdd5AufqSSqeM2qN1xzybapC8G4wEGGkZwyTDt1v"
          },
          "inputAmount": {
            "type": "number",
            "description": "Количество входного токена",
            "example": 1.5,
            "minimum": 0
          },
          "priceRangePercent": {
            "type": "number",
            "description": "Процент ценового диапазона",
            "example": 2,
            "minimum": 0.1,
            "maximum": 100
          }
        },
        "required": [
          "baseMint",
          "quoteMint",
          "inputAmount",
          "priceRangePercent"
        ]
      },
      "ClosePositionDto": {
        "type": "object",
        "properties": {
          "nftMint": {
            "type": "string",
            "description": "NFT Mint адрес позиции",
            "example": "5HauEoJa6tcKBy1vofNAxnEDiMpC7H6V9L5TgLwFYNfd"
          }
        },
        "required": [
          "nftMint"
        ]
      },
      "IncreasePositionDto": {
        "type": "object",
        "properties": {
          "nftMint": {
            "type": "string",
            "description": "NFT Mint адрес позиции",
            "example": "5HauEoJa6tcKBy1vofNAxnEDiMpC7H6V9L5TgLwFYNfd"
          },
          "inputAmount": {
            "type": "number",
            "description": "Input Amount",
            "example": "1"
          }
        },
        "required": [
          "nftMint",
          "inputAmount"
        ]
      },
...
    }
  }
}